+++ packages/react-magma-dom/src/components/Tabs/Tabs.tsx
@@ -380,6 +380,9 @@ export const Tabs = React.forwardRef<HTMLDivElement, TabsProps & Orientation>(
       }
 
       if (newActiveTabIndex !== null) {
+        onChange &&
+          typeof onChange === 'function' &&
+          onChange(newActiveTabIndex);
         setActiveTabIndex(newActiveTabIndex);
         (buttonRefArray.current[newActiveTabIndex]
           .current as HTMLButtonElement).focus();