+++ packages/react-magma-dom/src/i18n/interface.ts
@@ -141,7 +141,7 @@ export interface I18nInterface {
     menuItemSelectedAriaLabel: string;
     toggleMenuAriaLabel: string;
   };
-  fileUploader: {
+  dropzone: {
     browseFiles: string;
     dragMessage: string;
     errors: {
@@ -151,6 +151,9 @@ export interface I18nInterface {
       };
     }
     files: string;
+    bytes: string;
+    deleteFile: string;
+    removeFile: string;
   };
   header: {
     navigationButtonLabel: string;