+++ packages/react-magma-dom/src/index.ts
@@ -208,6 +208,7 @@ export {
   TypographyContextVariant,
   TypographyVisualStyle,
 } from './components/Typography';
+
 export {
   VisuallyHidden,
   VisuallyHiddenProps,
@@ -233,9 +234,6 @@ export { InverseContext, useIsInverse } from './inverse';
 export * from './components/Accordion';
 export * from './components/Transition';
 export * from './components/Drawer';
-
 export * from './components/FormFieldContainer';
-
 export * from './components/Grid';
-
 export * from './components/Tag';