+++ website/react-magma-docs/src/components/CodeBlock/preview.tsx
@@ -10,6 +10,7 @@ import styled from '@emotion/styled';
 import { es, zhCN } from 'date-fns/locale';
 import * as HEADER from '@cengage-patterns/header';
 import * as CHARTS from '@react-magma/charts';
+import * as DROPZONE from '@react-magma/dropzone';
 
 import { CodeBlockContext } from './context';
 
@@ -60,7 +61,17 @@ export const Preview = ({ ...props }: PreviewProps) => {
     <PreviewContainer isInverse={isInverse} noBorder={props.noBorder}>
       <LiveProvider
         code={preview.code}
-        scope={{ es, zhCN, styled, ...MAGMA, ...ICONS, ICONS, ...HEADER, ...SCHEMA }}
+        scope={{
+          es,
+          zhCN,
+          ...MAGMA,
+          ...ICONS,
+          ICONS,
+          ...HEADER,
+          ...SCHEMA,
+          ...DROPZONE,
+          ...CHARTS,
+        }}
         noInline
         transformCode={transformCode}
       >